CT-0007 Fix styles
