+++ index.js
@@ -6,4 +6,5 @@ import { AppRegistry } from 'react-native';
 import { name as appName } from './app.json';
 import App from './App';
 
+console.disableYellowBox = true;
 AppRegistry.registerComponent(appName, () => App);
