+++ src/components/Category.tsx
@@ -1,6 +1,6 @@
 import Product from '@components/Product';
 import { commonStyles } from '@styles/common';
-import { styles } from '@styles/productCategory';
+import { styles } from '@styles/category';
 import React from 'react';
 import { ScrollView, Text, View } from 'react-native';
 import { ICategory, IProduct } from 'types';
@@ -13,7 +13,7 @@ const byOrder = (a: IProduct, b: IProduct): number => {
   return a.order - b.order;
 };
 
-export const ProductCategory = (props: Props) => {
+export const Category = (props: Props) => {
   const { products, name } = props.cat;
   return (
     <View style={{ ...styles['product-category'] }}>
@@ -21,8 +21,7 @@ export const ProductCategory = (props: Props) => {
         style={{
           ...commonStyles.font,
           ...commonStyles['font--bolder'],
-        }}
-      >
+        }}>
         {name}
       </Text>
       <ScrollView style={styles['product-category__scroll']} bounces horizontal>
