+++ src/pages/Home/Market.tsx
@@ -1,44 +1,21 @@
-import { AVATAR_URI, IMAGE_URI } from '@common/constants';
-import { Price } from '@common/price';
-import { range } from '@common/range';
-import { ProductCategory } from '@components/ProductCategory';
-import { commonStyles } from '@styles/common';
-import React from 'react';
+import { Category } from '@components/Category';
+import { getProductData } from '@side-effects/server';
+import { styles } from '@styles/market';
+import React, { useEffect, useState } from 'react';
 import { View } from 'react-native';
-import { IProduct } from 'types';
-
-const mkProduct = (category: string) => (n: number): IProduct => ({
-  seller: {
-    imageUri: `${AVATAR_URI}+${n}`,
-  },
-  price: Price.of(Math.round(Math.round(1000)))
-              .discountBy(11)
-              .amount(),
-  title: `Product ${n}`,
-  description: `Description ${n}`,
-  picUri: `${IMAGE_URI}${category}+${n}`,
-  order: Math.round(Math.random() * 10),
-});
-
-const mkCategory = (name: string) => ({
-  name,
-  products: range(10).map(mkProduct(name)),
-});
+import { ICategory } from 'types';
 
 const Market = () => {
+  const [categories, setCategories] = useState<ICategory[]>([]);
+
+  useEffect(() => {
+    getProductData().then(setCategories);
+  }, []);
   return (
-    <View
-      style={{
-        ...commonStyles.avoidStatusBar,
-        ...commonStyles.margin,
-        ...commonStyles['full-height'],
-        overflow: 'hidden',
-        width: '97%',
-      }}
-    >
-      <ProductCategory cat={mkCategory('Ships')} />
-      <ProductCategory cat={mkCategory('Planes')} />
-      <ProductCategory cat={mkCategory('Cars')} />
+    <View style={styles.market}>
+      {categories.map(cat => (
+        <Category key={cat.name} cat={cat} />
+      ))}
     </View>
   );
 };
