+++ src/styles/category.ts
@@ -2,12 +2,13 @@ import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
   'product-category': {
-    height: '29%',
     marginBottom: 5,
+    flex: 1,
   },
   'product-category__scroll': {
     marginTop: 3,
     display: 'flex',
+    height: '100%',
     flexDirection: 'row',
   },
 });
