+++ tsconfig.json
@@ -24,6 +24,9 @@
             "@pages/*": [
                 "src/pages/*"
             ],
+            "@side-effects/*": [
+                "src/side-effects/*"
+            ],
             "types": [
                 "src/types/index.d.ts"
             ]
